fix transaction history race conditions and simplify API models
- Fix pagination race conditions by canceling in-flight requests on filter change
- Include existing data in LoadingMore state to prevent UI flicker
- Use Locale.ROOT for ISO 8601 date parsing
- Remove tokenSymbol from Transaction (not returned by API)
- Remove unused TOKEN_ADDRESS_TO_SYMBOL mapping

+++ product/pos/src/main/kotlin/com/walletconnect/pos/api/ApiModels.kt
@@ -132,6 +132,6 @@ internal data class TransactionStatsDto(
 
 @JsonClass(generateAdapter = true)
 internal data class TotalRevenueDto(
-    @param:Json(name = "amount") val amount: Int,
+    @param:Json(name = "amount") val amount: Double,
     @param:Json(name = "currency") val currency: String
 )
